video-recorder/requirements.txt
aiohttp==3.9.1
pyyaml==6.0.1
psutil==5.9.6
ntplib==0.4.0
asyncio==3.4.3